agent/requirements.txt
google-adk
python-dotenv
composio_openai
fastapi
requests